site/requirements.txt
markdown
pystache
PyYAML
watchdog
Pillow